ARM: S5PC11X: CPUFREQ: Fix power suspend cpu frequency
Power suspend often crashes the device if the cpu frequency is not
equal to SLEEP_FREQ. If the current cpufreq policy does not include
SLEEP_FREQ, cpufreq_driver_target() does not change the cpu frequency
for power suspend, resulting in crashes. This patch modifies the
current cpufreq policy to include SLEEP_FREQ for suspend, and
restores the cpufreq policy after suspend.

Change-Id: I45ac6f959365870c3ef05e31a8befd52c1231f42
Signed-off-by: Michael Huang <mike.g.huang@gmail.com>
Signed-off-by: nubecoder <nubecoder@gmail.com>